refactor(account-transfer): extract TransferQRScanner logic to MVVM view model
Move device registration, scan handling, and state management into
useTransferQRScannerViewModel hook, leaving the screen as a pure view.
